uintptr_t pour printptr et ptrhexa

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: bguillau <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 17:09:54 by bguillau          #+#    #+#             */
-/*   Updated: 2022/12/23 13:16:25 by bguillau         ###   ########.fr       */
+/*   Updated: 2022/12/23 14:37:27 by bguillau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@
 # include <stdarg.h>
 # include <unistd.h>
 # include <stdlib.h>
+# include <stdint.h>
 
 int		ft_printf(const char *str, ...);
 int		print_char(va_list init);
@@ -27,7 +28,7 @@ int		print_str(va_list init);
 int		print_ptr(va_list init);
 int		print_deci(va_list init);
 int		print_unsigned(va_list init);
-int		ptr_hexa(unsigned long long nbr, int *len, char c);
+int		ptr_hexa(uintptr_t nbr, int *len, char c);
 int		print_hexa(va_list init, char xX);
 void	ft_putunbr_fd(unsigned int n, int fd);
